Models and variables U

+++ back/src/main/scala/phystech/responses/ChangeModelResponse.scala
@@ -1,3 +1,3 @@
 package phystech.responses
 
-final case class ChangeModelResponse(modelId: String, parentModelId: String, version: Int)
+final case class ChangeModelResponse(modelId: String, parentModelId: String, version: Long)
